Introduce pokecount parameter for batch fetching

+++ app/Manage/fetchPokemon/FetchPokemon.onSubmit.php
@@ -21,41 +21,47 @@ if (isset($_GET['action'])) {
 
     switch ($_GET['action']) {
         case 'Fetch':
-            // fetch pokemon: insert pokename, poketype; update pokedex
-            $pokemona = json_decode($api->pokemon($_GET['pokeid']),true);
-            unset($pokemona['stats']);
-            unset($pokemona['abilities']);
-            unset($pokemona['moves']);
-            unset($pokemona['held_items']);
-            unset($pokemona['game_indices']);
-            $pokename = [
-                'pokeid'        => $pokemona['id'],
-                'pokename'      => $pokemona['name'],
-                'pokename_ru'   => '',
-                // 'sprites'   => $pokemona['sprites'],
-            ];
-            $poketype = [];
-            foreach ($pokemona['types'] as $i=>$a)
-                $poketype[]=[$pokename['pokeid'],$a['type']['name']];
+            if (!isset($_GET['pokeid'])) break;
+            if (!isset($_GET['pokecount'])) $_GET['pokecount']=1;
+            for ($pokecount=0;$pokecount<$_GET['pokecount'];$pokecount++) {
+                // fetch pokemon: insert pokename, poketype; update pokedex
+                $pokemona = json_decode($api->pokemon($_GET['pokeid']),true);
+                unset($pokemona['stats']);
+                unset($pokemona['abilities']);
+                unset($pokemona['moves']);
+                unset($pokemona['held_items']);
+                unset($pokemona['game_indices']);
+                $pokename = [
+                    'pokeid'        => $pokemona['id'],
+                    'pokename'      => $pokemona['name'],
+                    'pokename_ru'   => '',
+                    // 'sprites'   => $pokemona['sprites'],
+                ];
+                $poketype = [];
+                foreach ($pokemona['types'] as $i=>$a)
+                    $poketype[]=[$pokename['pokeid'],$a['type']['name']];
 
-            // create records
-            if ($tbPokename->insert($pokename)) {
-                if ($tbPoketype->insert($poketype)) {
-                    // update pokedex
-                    if (!$tbPokedex->update(['localdata'=>1],false,'pokeid='.$pokename['pokeid'])) {
-                        logMessage('FetchPokemon','pokedex update: '.sqlError(),'danger');
+                // create records
+                if ($tbPokename->insert($pokename)) {
+                    if ($tbPoketype->insert($poketype)) {
+                        // update pokedex
+                        if (!$tbPokedex->update(['localdata'=>1],false,'pokeid='.$pokename['pokeid'])) {
+                            logMessage('FetchPokemon','pokedex update: '.sqlError(),'danger');
+                        }
+                    } else {
+                        logMessage('FetchPokemon','poketype: '.sqlError(),'danger');
+                        $tbPokename->delete('pokeid='.$pokename['pokeid']);
                     }
                 } else {
-                    logMessage('FetchPokemon','poketype: '.sqlError(),'danger');
-                    $tbPokename->delete('pokeid='.$pokename['pokeid']);
+                    logMessage('FetchPokemon','pokename: '.sqlError(),'danger');
                 }
-            } else {
-                logMessage('FetchPokemon','pokename: '.sqlError(),'danger');
+                // log results
+                foreach ($pokemona['types'] as $i=>$a)
+                    $pokename['type'][]=$a['type']['name'];
+                logMessage('FetchPokemon','Fetched: '.varExport($pokename));
+                // next
+                $_GET['pokeid']++;
             }
-            // log results
-            foreach ($pokemona['types'] as $i=>$a)
-                $pokename['type'][]=$a['type']['name'];
-            logMessage('FetchPokemon','Fetched: '.varExport($pokename));
             break;
     }
 }
